% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigpro.R
\name{decomposition}
\alias{decomposition}
\title{decomposition}
\usage{
decomposition(signatures, activities, samples, output,
  mutation_type = "96", genome_build = "GRCh37", verbose = F)
}
\arguments{
\item{signatures}{A string. Path to a  tab delimited file that contains the signaure table where the rows are mutation types and colunms are signature IDs.}

\item{activities}{A string. Path to a tab delimilted file that contains the activity table where the rows are sample IDs and colunms are signature IDs.}

\item{samples}{A string. Path to a tab delimilted file that contains the activity table where the rows are mutation types and colunms are sample IDs.}

\item{output}{A string. Path to the output folder.}

\item{mutation_type}{A string. The context type. Example: "96", "192", "1536", "6144", "INDEL", "DINUC". The default value is "96".}

\item{genome_build}{A string. The genome type. Example: "GRCh37", "GRCh38", "mm9", "mm10". The default value is "GRCh37"}

\item{verbose}{Boolean. Prints statements. Default value is False.}
}
\value{
A folder with decomposition results The files below will be generated in the output folder: \cr
   \cr
   1.comparison_with_global_ID_signatures.csv \cr
   3.Decomposed_Solution_Samples_stats.txt \cr
   4.Decomposed_Solution_Signatures.txt \cr
   5.decomposition_logfile.txt \cr
   6.Mutation_Probabilities.txt \cr
   7.Signature_assaignment_logfile.txt \cr
   8.Signature_plot[MutatutionContext]_plots_Decomposed_Solution.pdf
}
\description{
Decomposes the De Novo Signatures to Cosmic Signatures
}
