% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigpro.R
\name{importdata}
\alias{importdata}
\title{Title}
\usage{
importdata(datatype)
}
\arguments{
\item{datatype}{The type of example data. There are two types: 1."table", 2."vcf"}
}
\value{
The path of the explample data
}
\description{
Title
}
