% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigpro.R
\name{sigProfilerExtractor}
\alias{sigProfilerExtractor}
\title{SigprofilerExtractor Function}
\usage{
sigProfilerExtractor(input_type, output, input, refgen = "GRCh37",
  genome_build = "GRCh37", minsigs = 1, maxsigs = 3,
  replicates = 5, mtype = c("96,DINUCS,ID"), init = "random",
  exome = F, cpu = -1)
}
\arguments{
\item{input_type}{The type of input}

\item{output}{The name/path of the result folder}

\item{input}{Input file or folder}

\item{refgen}{Reference genome}

\item{genome_build}{Genome opportunity}

\item{minsigs}{Minimum number of signatures to be extracted}

\item{maxsigs}{Maximum number of signatures to be extracted}

\item{replicates}{Number of replication of signature extraction}

\item{mtype}{Context type}

\item{init}{NMF initialization}

\item{exome}{Exome true of False}

\item{cpu}{Number of CPUs to be used during multiprocessing}
}
\value{
a folder with results
}
\description{
SigprofilerExtractor Function
}
