% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigpro.R
\name{sigprofilerextractor}
\alias{sigprofilerextractor}
\title{sigprofilerextractor}
\usage{
sigprofilerextractor(input_type, output, inputdata, refgen = "GRCh37",
  genome_build = "GRCh37", minsigs = 1, maxsigs = 3,
  replicates = 5, mtype = c("96,DINUCS,ID"), init = "random",
  exome = F, cpu = -1)
}
\arguments{
\item{input_type}{input_type: A string. Type of input. The type of input should be one of the following:\cr
----- "vcf": used for vcf format inputs.\cr
----- "table": used for table format inputs using a tab seperated file.\cr}

\item{output}{A string. The name of the output folder. The output folder will be generated in the current working directory.\cr}

\item{inputdata}{A string. Name of the input folder (in case of "vcf" type input) or the input file (in case of "table"  type input).\cr
The project file or folder should be inside the current working directory.\cr
For the "vcf" type input,the project has to be a folder which will contain the vcf files in vcf format or text formats.\cr
The "text"type projects have to be a file.}

\item{refgen}{A string, optional. The name of the reference genome. The default reference genome is "GRCh37". This parameter is applicable only if the input_type is "vcf".\cr}

\item{minsigs}{A positive integer, optional. The minimum number of signatures to be extracted. The default value is 1}

\item{maxsigs}{A positive integer, optional. The maximum number of signatures to be extracted. The default value is 10}

\item{replicates}{A positive integer, optional. The number of iteration to be performed to extract each number signature. The default value is 8}

\item{mtype}{A  strings, optional. The items in the list defines the mutational contexts to be considered to extract the signatures. The default value is "96,DINUCS,ID",\cr
where "96" is the SBS96 context, "DINUC" is the DINULEOTIDE context and ID is INDEL context. Each context should be written in a single string separated by "," without any space.}

\item{init}{A string. Represents the NMF initialization. Options are "random", "nndsvdar" and "nndsvd"}

\item{exome}{Boolean, optional. Defines if the exomes will be extracted. The default value is "False".}

\item{cpu}{An integer, optional. The number of processors to be used to extract the signatures. The default value is -1 which will use all available processors.}
}
\value{
a folder with results
}
\description{
Extracts De Novo signature from genomes and Decomposes the De Novo Signatures to Cosmic Signatures
}
